Refactor dio file
